Use shared EarliestScheduler in automations service
